requirements.txt
asgiref==3.10.0
Django==4.2.25
django-cors-headers==4.9.0
django-filter==25.1
django-sslserver==0.22
djangorestframework==3.16.1
elasticsearch==8.11.0
elasticsearch-dsl==8.11.0
gunicorn==23.0.0
packaging==25.0
pillow==11.3.0
psycopg2-binary==2.9.11
sqlparse==0.5.3
typing_extensions==4.15.0
whitenoise==6.11.0
